vimeoapi/pyproject.toml
[tool.poetry]
name = "vimeorganizer"
version = "0.1.0"
description = ""
authors = ["trop3n <jasondkimm@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pyvimeo = "^1.1.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
